r.li: fix memory handling (memory leak in avl_to_array function)
avl_to_array function was allocating one by one structures which were already allocated by callers. Callers were not freeing the memory allocated for structures. Also avl_to_array function was not respecting the size and type of the passed array, so the dereferencing was probbaly not working correctly. The type of array passed to the function is now AVL_table, not the pointer to it, because this is already a pointer.

The influenced modules now should run faster, for small resolutions, and with better (correct) results.

The naming of AVL_table and AVL_tableRow is still strange.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@61812 15284696-431f-4ddb-bdfa-cd5b030d7da7