do not use tabs which are 8 spaces wide when indent level is 4 spaces
Indent script was creating source code with 4 spaces for indent level 1,
1 tab for indent level 2, 1 tab and 4 spaces for indent level 3, 2 tabs
for indent level 4 etc. This is style which is hard to do manually
and cannot be easily guessed and supported by a text editor because it
mixes spaces and tabs
(http://lists.osgeo.org/pipermail/grass-dev/2014-September/070653.html).

The indent rules are still the same, each level must be in multiplies
of 4 spaces (columns) as specified in
http://lists.osgeo.org/pipermail/grass-dev/2014-August/070497.html.

No backwards compatibility is needed because just small portion of the
current code has no diff after applying the (original version of) indent
script. Now tested with v.external, i.segment and r.slope.aspect.

The -ut (--use-tabs) was in the list at
https://trac.osgeo.org/grass/wiki/Submitting/C.
But it was not in the list at
http://grasswiki.osgeo.org/w/index.php?title=Development&oldid=21058#Explanation_of_C_indentation_rules.
And tabs are explicitly forbidden in instructions for Emacs users at
http://grasswiki.osgeo.org/w/index.php?title=GRASS_Programming_Howto&oldid=20573#Emacs.

See also #1663 (Tabs in C source code) and grass-dev Tabs and spaces in C code
(http://lists.osgeo.org/pipermail/grass-dev/2014-August/070484.html).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@64658 15284696-431f-4ddb-bdfa-cd5b030d7da7