init: escape backslash in prompt string
This avoids anomalous-backslash-in-string warning (W1401) in Pylint. It is just a warning because the unknown escape sequence is ignored and thus the backslash is preserved for bash. The other solution would be to use raw strings but this complicates usage of newline character. BTW, backslash lowercase w means 'current absolute path' for bash.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@65347 15284696-431f-4ddb-bdfa-cd5b030d7da7