init: end immediately when GUI not available
When GUI to select the Location and Mapset is not available (usually because of missing wxPython),
end execution right away instead of trying the text mode. The text mode was not starting anyway when
GRASS was just compiled. When GRASS was installed, it got into a loop with 'Hit RETURN to continue'
(for unknown reasons). User anyway had not specified the Mapset, so there is no point in selecting
the last one used.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@65849 15284696-431f-4ddb-bdfa-cd5b030d7da7