g.extension: rename option svnurl to url
The name url is not 100% friendly for all when used for local files but source
(with unclear meaning) and input (usually an actual file to process) don't
seem better.

Also clean up in the GUI wrapper, document tests and fix bug in listing from HTML
(introduced in https://trac.osgeo.org/grass/changeset/65868).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@65869 15284696-431f-4ddb-bdfa-cd5b030d7da7