v.distance: use bbox in 2D and 3D accordingly (see #2734)
Check if we are considering z and use appropriate point in box function. This should
fix #2734 where point inside an area is classified as outsie of that area when point
3D but the area is 2D. The expected behavior is reduction to 2D.

The test works with this commit but its 3D part fails with code prior to this commit.

This commit depens on r66459 which adds Vect_point_in_box_2d() function.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@66460 15284696-431f-4ddb-bdfa-cd5b030d7da7