v.vol.rst: remove unused write to array
Causes segfaut when number of points is much smaller
then number of columns (valgrind: invalid write of
size 8).

Is seems that A is not used anymore in these loops
and for next run it is initialized again.

The index used is completely out of range
(number of columns versus number of points).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@68112 15284696-431f-4ddb-bdfa-cd5b030d7da7