pyedit: do not use onDone because it is not executed when GUI is invoked (fixes #2997)
Now you can run again even if it is already running
which is the same behavior as for normal module calls.

It also uses the same temporary file once it was created
to avoid the need to delete it all the time. It is deleted
when file is actually saved.

Also remove suppression of the GUI interface because we
actually want to get GUI (the parameter does not work, r68395).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@68396 15284696-431f-4ddb-bdfa-cd5b030d7da7