init: PEP8 the example
In case we are trying to set good example for the beginners by the modern print() in https://trac.osgeo.org/grass/changeset/70020,
let's set good example also by using PEP8 (W293, E251 and potential E211).

But the whole change may not be that friendly to beginners because
G does not support Python 3 yet and print() here is still not a function,
so using it as the Python 3 one fails.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@70027 15284696-431f-4ddb-bdfa-cd5b030d7da7