rst: add raster and gmath to libraries
Also reorder libraries to more common order.
Add dependencies synced to libs.
Leave math in libs as in other modules
although it likely should be in extra or just get it from link command.

This is meant to fix compilation on Mac where some of the libs are
in EXTRA_LIBS which is reasigned.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@70334 15284696-431f-4ddb-bdfa-cd5b030d7da7