v.edit: Add extend,extendstart,extendend tools with -p flag
These tools are similar to connect, but connect is mainly intended to connect
two lines when the first line can be extended onto the second line. If the
second line is short and extending the first line doesn't intersect the second
one, the minimum distance is not a normal distance and the connect tool can
change the shape of the second line by moving an existing vertex before
connecting two lines. Not sure if that was intended. The extend tools try to
extend both lines until they intersect so that the existing shape doesn't
change. This is important especially for hydrologic analysis such as the
longest flow path on top of the drainage raster map. Since these tools extend
lines to connect them, they don't work for parallel lines by default. Use the
-p flag to add a new node to the first line to connect both lines.



git-svn-id: https://svn.osgeo.org/grass/grass/trunk@70562 15284696-431f-4ddb-bdfa-cd5b030d7da7