start box (square) symbol in the middle of straight line
The original symbol has bad top left corner when line width is big
because the line ends in the middle of the other line width.
This makes the connection between start and end lines in the
part where the shape is straight so there is nothing special happening
when one line ends and the other begins.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@70799 15284696-431f-4ddb-bdfa-cd5b030d7da7