wxGUI/mapdisp: change location@mapset to location/mapset in title
The syntax location@mapset uses @ which has different meaning in GRASS.
Location: at the beginning was not applicable anymore (was OK when there
was no mapset in the title.
Version removed from the title to make it shorter and consistent with other
windows such as Animation or Histogram Tool. Version is in Layer Manager
window title.
API standardized and its usage clarified (but still not ideal).
API used consitently so 'on map display rename' mapset is included as well
and title code is not duplicated.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@71301 15284696-431f-4ddb-bdfa-cd5b030d7da7