init: -c not needed for --tmp-location (closes https://trac.osgeo.org/grass/ticket/3585)
--tmp-location required -c flag because -c was associated with EPSG specification.
The new approach is that EPSG is a value of whatever flag/option which needs
EPSG or equivalent.

No changes in the code. Parsing just takes any ARG (non-flag) available.
We check if EPSG was provided and that mapset was not provided (i.e. exactly
one ARG was provided). It is no longer possible to use --tmp-location without
ARG to create an XY location.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73096 15284696-431f-4ddb-bdfa-cd5b030d7da7