init: advertise the double dash option in help text
Both are supported already for a long time, --text is the more standard
(comparing to -text) plus it is the one used with long flags for modules
(e.g. --verbose). For now, one dash is still supported, so not warning
the user when -text etc are used.

See also wiki:Grass8Planning#grassexecutable and comment:2:ticket:3585.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73100 15284696-431f-4ddb-bdfa-cd5b030d7da7