wxGUI/core: replace grass.script.tempfile by tempfile.NamedTemporaryFile (fixes https://trac.osgeo.org/grass/ticket/3637)
Switches from mapset temporary directory to system temporary
directory which was previously set up for the session for
all GUI rendering files.

This also refactors and unifies handling of the temporary files related
to rendering, although on error and when removing/deleting the layers
removing of the files happens in the same way as before.

All files are now using NamedTemporaryFile instead of some using
g.tempfile and some NamedTemporaryFile. This means that when mapset
is switched, files are present in the (correct) session directory,
not incorrect mapset directory and then not deleted later.
(Deleting these files could cause rendering errors, while not deleting
them would lead to accumulating of files in mapset temporary directory).

See also:
 * https://trac.osgeo.org/grass/changeset/28605 which adds glob for one delete when core.utils.GetTempfile(),
   i.e., g.tempfile was still used (the glob part is still there)
 * https://trac.osgeo.org/grass/changeset/56444 which adds mkstemp() for some of the files
 * https://trac.osgeo.org/grass/changeset/60947 attempt to fix cleaning of deleting temporary files (https://trac.osgeo.org/grass/ticket/560)
   which changes mkstemp to NamedTemporaryFile (still used now)
 * https://trac.osgeo.org/grass/changeset/69085 which checked legend row file (legrow) name of last character
   being digit (isdigit()) which was true for files from g.tempfile
   but does not seem to have a reason (removed now)
 * https://trac.osgeo.org/grass/ticket/3635 which investigates a suspicious cleanups of mapset temporary
   directory which at this point seemed to be cleaning rendering files
   from GUI (https://trac.osgeo.org/grass/changeset/37863, https://trac.osgeo.org/grass/changeset/21048)


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73334 15284696-431f-4ddb-bdfa-cd5b030d7da7