init: remove lock from the current mapset (fixes https://trac.osgeo.org/grass/ticket/3631)
Do not remove lock based on the path to the initially locked file,
but instead use the rc file point to by GISRC (assumed stable within this proccess)
to find out the current mapset and remove lock file there.

Now a sequence start in mapset A, switch to B, exit removes all lock files.
Before, only the A lock file was removed, specifically the lock file was left
in the mapset exited from.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73335 15284696-431f-4ddb-bdfa-cd5b030d7da7