init: clean mapset only after locking it (fixes https://trac.osgeo.org/grass/ticket/613, see https://trac.osgeo.org/grass/ticket/1286)
Only after we set the user-requested mapset as current (GISRC) and lock it,
we can do the startup cleaning using clean_temp (to clean after previous sessions
if needed).

The cleaning was in place at least since https://trac.osgeo.org/grass/changeset/11260 but it got moved to a wrong place some
time after that but before r37863.


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73342 15284696-431f-4ddb-bdfa-cd5b030d7da7