wxGUI: save database, location and mapset to workspace and switch it when opening it (closes https://trac.osgeo.org/grass/ticket/3575)
It is no longer possible to just open workspace in a different mapset
than the one where it was saved. However, when session element is not
present (deleted or older gxw file) no mapset switching is performed.
Older versions of GRASS GIS simply ignore the session element.

An error message is given when mapset does not exist and options to
proceed or not are given with some details (can be improved by
diagnostics from grass.py if moved to library).


git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73348 15284696-431f-4ddb-bdfa-cd5b030d7da7